refactor: Move AST into its own module under src/txxt_nano/ast/
Previously, AST definitions were scattered across multiple files in the parser
module (ast.rs, source_location.rs, ast_position.rs). This refactoring
organizes the AST into a dedicated module with clear separation of concerns:

- ast/span.rs: Position and Span types for source code tracking
- ast/node.rs: All AST node definitions and trait implementations
- ast/position.rs: SourceLocation utility for byte-to-position conversion
- ast/error.rs: Error types for AST operations
- ast/lookup/: Position-based AST node lookup functionality

All parser and format modules have been updated to import from the new
ast:: module instead of parser::. This change provides better organization
and clearer API boundaries.

All tests pass successfully (175 passed).
