+++ src/txxt_nano.rs
@@ -1,5 +1,6 @@
 //! Main module for txxt-nano library functionality
 
+pub mod ast;
 pub mod formats;
 pub mod lexer;
 pub mod parser;
