+++ src/txxt_nano/ast/lookup/by_position.rs
@@ -3,7 +3,8 @@
 //! This module provides functionality to find and display AST nodes at a given
 //! position (line:column) in the source code.
 
-use super::ast::{AstNode, ContentItem, Document, Position, Span};
+use super::super::node::{AstNode, ContentItem, Document};
+use super::super::span::{Position, Span};
 use crate::txxt_nano::processor::ProcessingError;
 use std::collections::HashMap;
 
@@ -163,7 +164,7 @@ fn get_content_item_label(item: &ContentItem) -> Option<String> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::txxt_nano::parser::{Document as DocumentType, Paragraph};
+    use crate::txxt_nano::ast::node::{Document as DocumentType, Paragraph};
 
     #[test]
     fn test_parse_position_valid() {
