+++ src/txxt_nano/ast/lookup/mod.rs
@@ -1,0 +1,7 @@
+//! AST node lookup functionality
+//!
+//! This module provides utilities for finding and querying AST nodes by position.
+
+pub mod by_position;
+
+pub use by_position::format_at_position;
