+++ src/txxt_nano/formats/tag/tag.rs
@@ -21,7 +21,7 @@
 //! </children></session>
 //! ```
 
-use txxt_ast::{Container, ContentItem, Document};
+use crate::txxt_nano::ast::{Container, ContentItem, Document};
 
 /// Serialize a document to AST tag format
 pub fn serialize_document(doc: &Document) -> String {
@@ -179,7 +179,7 @@ fn escape_xml(text: &str) -> String {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use txxt_ast::{Paragraph, Session};
+    use crate::txxt_nano::ast::{Paragraph, Session};
 
     #[test]
     fn test_serialize_simple_paragraph() {
