+++ src/txxt_nano/parser/tests.rs
@@ -1,3 +1,4 @@
+use crate::txxt_nano::ast::ContentItem;
 use crate::txxt_nano::lexer::{lex, lex_with_spans, Token};
 use crate::txxt_nano::parser::api::parse;
 use crate::txxt_nano::parser::ast_conversion::convert_paragraph;
